Allow app to serve both html and erb files.

+++ app.rb
@@ -1,12 +1,23 @@
+require 'erb'
+
 class App
   def call(env)
     [
         200,
         {
-            'Content-Type'  => 'text/html',
+            'Content-Type' => 'text/html',
             'Cache-Control' => 'public, max-age=86400'
         },
-        File.open('public/index.html', File::RDONLY)
+        body
     ]
+  end
+
+  private
+  def body
+    if File.exist?('public/index.html')
+      File.open('public/index.html', File::RDONLY)
+    else
+      [ERB.new(File.open('public/index.html.erb', File::RDONLY).read).result]
+    end
   end
 end
